Refactor v2 parser conditions
